feat(dynamic-relays): move test helper into a separate file
